Update repository interfaces to use Integer instead of Long for IDs

+++ src/main/java/CENG453/group18/repository/GameBoardRepository.java
@@ -3,6 +3,6 @@ package CENG453.group18.repository;
 import CENG453.group18.entity.GameBoard;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GameBoardRepository extends JpaRepository<GameBoard, Long> {
-    public GameBoard getGameBoardById(int id);
+public interface GameBoardRepository extends JpaRepository<GameBoard, Integer> {
+    public GameBoard getGameBoardById(Integer id);
 }
